requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy<2.0.0
scipy>=1.10.0
PyYAML>=6.0
tqdm>=4.65.0
gdown>=4.7.0
pathlib
